资源的注册添加，删除，查询

+++ src/main/java/group/fire_monitor/service/impl/ResourceServiceImpl.java
@@ -1,16 +1,20 @@
 package group.fire_monitor.service.impl;
 
+import com.auth0.jwt.JWT;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.jcraft.jsch.JSch;
 import com.jcraft.jsch.JSchException;
 import com.jcraft.jsch.Session;
+import group.fire_monitor.mapper.RelationGroupResourceMapper;
+import group.fire_monitor.mapper.RelationGroupUserMapper;
 import group.fire_monitor.mapper.ResourceMapper;
-import group.fire_monitor.mapper.UserMapper;
+import group.fire_monitor.pojo.RelationGroupResource;
+import group.fire_monitor.pojo.RelationGroupUser;
 import group.fire_monitor.pojo.Resource;
-import group.fire_monitor.pojo.User;
+import group.fire_monitor.pojo.form.ResourceCreateForm;
 import group.fire_monitor.service.ResourceService;
-import group.fire_monitor.service.UserService;
+import group.fire_monitor.util.JWTUtil;
+import group.fire_monitor.util.enums.PermissionLevelEnum;
 import group.fire_monitor.util.response.UniversalResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,17 +23,20 @@ import org.springframework.stereotype.Service;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
-import java.net.InetAddress;
 import java.net.URL;
 import java.sql.Timestamp;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Service
 public class ResourceServiceImpl implements ResourceService {
     @Autowired
     ResourceMapper resourceMapper;
+    @Autowired
+    RelationGroupResourceMapper relationGroupResourceMapper;
+    @Autowired
+    RelationGroupUserMapper relationGroupUserMapper;
 
     @Override
     public UniversalResponse<?> testPing(Resource resource) {
@@ -59,15 +66,25 @@ public class ResourceServiceImpl implements ResourceService {
     }
 
     @Override
-    public UniversalResponse<?> addServer(Resource resource) {
+    public UniversalResponse<?> addServer(ResourceCreateForm resourceCreateForm) {
         QueryWrapper<Resource> wrapper=new QueryWrapper<>();
-        wrapper.eq("resource_ip",resource.getResourceIp());
+        wrapper.eq("resource_ip",resourceCreateForm.getResource().getResourceIp());
         List<Resource> resources=resourceMapper.selectList(wrapper);
         if(!resources.isEmpty())return new UniversalResponse<>(500,"当前ip服务器已经被监控");
         try{
-            resource.setResourceType("server");
-            resource.setAddedTime(new Timestamp(System.currentTimeMillis()));
-            resourceMapper.insert(resource);
+            //注册资源表
+            resourceCreateForm.getResource().setResourceType("server");
+            resourceCreateForm.getResource().setAddedTime(new Timestamp(System.currentTimeMillis()));
+            resourceCreateForm.getResource().setResourceCreaterId(JWTUtil.getCurrentUser().getId());
+            resourceMapper.insert(resourceCreateForm.getResource());
+            //注册资源访问权限表
+            for(Integer groupId:resourceCreateForm.getGroupIdList()){
+                RelationGroupResource relationGroupResource=new RelationGroupResource();
+                relationGroupResource.setGroupId(groupId);
+                relationGroupResource.setResourceId(resourceCreateForm.getResource().getId());
+                relationGroupResourceMapper.insert(relationGroupResource);
+            }
+
             //TODO:JAVA API实现注册和自动重启(需要监测有没有安装exporter)
             return new UniversalResponse<>().success();
         }catch (Exception e){
@@ -78,8 +95,11 @@ public class ResourceServiceImpl implements ResourceService {
 
     @Override
     public UniversalResponse<?> deleteServer(Integer resourceId) {
+
         try{
+            checkPermission(resourceId);
             resourceMapper.deleteById(resourceId);
+            relationGroupResourceMapper.delete(new QueryWrapper<RelationGroupResource>().eq("resource_id",resourceId));
             //TODO:JAVA API实现注册和自动重启
             return new UniversalResponse<>().success();
         }catch (Exception e){
@@ -100,6 +120,11 @@ public class ResourceServiceImpl implements ResourceService {
                     .like("resource_name",str)
                     .like("resource_description",str);
             List<Resource> resourceList= resourceMapper.selectList(wrapper);
+
+            //TODO:查询时加入权限筛选？
+//            List<Resource> filteredResources = resourceList.stream()
+//                    .filter(resource -> checkPermission(resourceList.stream().map(Resource::getId).collect(Collectors.toList())).contains(resource.getId())) // 筛选条件
+//                    .collect(Collectors.toList());
             return new UniversalResponse<>().success(resourceList);
         }catch (Exception e){
             return new UniversalResponse<>(500,e.getMessage());
@@ -108,15 +133,26 @@ public class ResourceServiceImpl implements ResourceService {
     }
 
     @Override
-    public UniversalResponse<?> addSoftware(Resource resource) {
+    public UniversalResponse<?> addSoftware(ResourceCreateForm resourceCreateForm) {
         QueryWrapper<Resource> wrapper=new QueryWrapper<>();
-        wrapper.eq("resource_ip",resource.getResourceIp());
+        wrapper.eq("resource_ip",resourceCreateForm.getResource().getResourceIp());
         List<Resource> resources=resourceMapper.selectList(wrapper);
         if(!resources.isEmpty())return new UniversalResponse<>(500,"当前ip服务器已经被监控");
         try{
-            resource.setResourceType("software");
-            resource.setAddedTime(new Timestamp(System.currentTimeMillis()));
-            resourceMapper.insert(resource);
+
+            //注册资源表
+            resourceCreateForm.getResource().setResourceCreaterId(JWTUtil.getCurrentUser().getId());
+            resourceCreateForm.getResource() .setResourceType("software");
+            resourceCreateForm.getResource() .setAddedTime(new Timestamp(System.currentTimeMillis()));
+            resourceMapper.insert(resourceCreateForm.getResource());
+            //注册资源访问权限表
+            for(Integer groupId:resourceCreateForm.getGroupIdList()){
+                RelationGroupResource relationGroupResource=new RelationGroupResource();
+                relationGroupResource.setGroupId(groupId);
+                relationGroupResource.setResourceId(resourceCreateForm.getResource().getId());
+                relationGroupResourceMapper.insert(relationGroupResource);
+            }
+
             //TODO:JAVA API实现注册(需要监测有没有安装exporter)
             return new UniversalResponse<>().success();
         }catch (Exception e){
@@ -126,8 +162,11 @@ public class ResourceServiceImpl implements ResourceService {
 
     @Override
     public UniversalResponse<?> deleteSoftware(Integer resourceId) {
+
         try{
+            checkPermission(resourceId);
             resourceMapper.deleteById(resourceId);
+            relationGroupResourceMapper.delete(new QueryWrapper<RelationGroupResource>().eq("resource_id",resourceId));
             //TODO:JAVA API实现解绑
             return new UniversalResponse<>().success();
         }catch (Exception e){
@@ -181,6 +220,10 @@ public class ResourceServiceImpl implements ResourceService {
                     .like("resource_type_second",str)
                     .like("resource_description",str);
             List<Resource> resourceList= resourceMapper.selectList(wrapper);
+            //TODO:查询时加入权限筛选？
+//            List<Resource> filteredResources = resourceList.stream()
+//                    .filter(resource -> checkPermission(resourceList.stream().map(Resource::getId).collect(Collectors.toList())).contains(resource.getId())) // 筛选条件
+//                    .collect(Collectors.toList());
             return new UniversalResponse<>().success(resourceList);
         }catch (Exception e){
             return new UniversalResponse<>(500,e.getMessage());
@@ -190,5 +233,26 @@ public class ResourceServiceImpl implements ResourceService {
     @Override
     public UniversalResponse<?> testExporter(Integer type) {
         return null;
+    }
+
+    @Override
+    public List<Integer> checkPermission(List<Integer> resourceIds) {
+         List<Integer> userInGroupIds=relationGroupUserMapper.selectList(new QueryWrapper<RelationGroupUser>().eq("user_id",JWTUtil.getCurrentUser().getId()))
+                .stream().map(RelationGroupUser::getGroup_id).collect(Collectors.toList());
+        return relationGroupResourceMapper.selectPermittedResourceId(userInGroupIds);
+    }
+
+    @Override
+    public void checkPermission(Integer resourceId) throws Exception {
+        if (Objects.equals(JWTUtil.getCurrentUser().getPermissionLevel(), PermissionLevelEnum.ADMIN.getPermissionLevel())) return;
+        List<RelationGroupResource> list= relationGroupResourceMapper.selectList(new QueryWrapper<RelationGroupResource>().eq("resource_id",resourceId));
+        List<Integer> groupIds=list.stream().map(RelationGroupResource::getGroupId).collect(Collectors.toList());
+        List<Integer> userInGroupIds=relationGroupUserMapper.selectList(new QueryWrapper<RelationGroupUser>().eq("user_id",JWTUtil.getCurrentUser().getId()))
+                .stream().map(RelationGroupUser::getGroup_id).collect(Collectors.toList());
+
+        List<Integer> intersection = groupIds.stream()
+                .filter(userInGroupIds::contains) // 筛选出 list2 中存在的元素
+                .collect(Collectors.toList());
+        if (intersection.isEmpty()) throw new Exception("无权限访问");
     }
 }
